Posix implementation
I felt like the need for threads could be interesting so as in previous branches I would want to implement it. The issue comes with all the different libraries and since I cannot make thread.h to work with G++ I decided to use either posix or boost::thread. Since pthreads seems to be the foundation I decided to go with that